# contract/api-spec.yaml
openapi: 3.0.0
info:
  title: Seedance Wizard API（国内版，无支付）
  version: 0.1.0
  description: |
    ComfyUI Seedance 向导的 Node.js 后端 API。
    负责：DeepSeek 导演 / Qwen VL 看图 / Flux 预览 / Seedance 出片代理 / API Key 管理 / 余额追踪。

    认证方式：所有 /api/* 路由（auth 除外）需在 Header 携带：
      Authorization: Bearer sk-seed-<token>
servers:
  - url: http://localhost:3000
    description: 本地开发
  - url: https://api.seedance.cn
    description: 国内生产环境（待定）

# ─────────────────────────────────────────────
# 通用数据结构
# ─────────────────────────────────────────────
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: sk-seed-<token>
  schemas:
    Error:
      type: object
      properties:
        code:    { type: string, example: "INSUFFICIENT_BALANCE" }
        message: { type: string, example: "余额不足，请充值" }
    ApiKey:
      type: object
      properties:
        id:          { type: string, format: uuid }
        name:        { type: string, example: "我的主力 Key" }
        key_prefix:  { type: string, example: "sk-seed-ab12" }
        is_active:   { type: boolean }
        created_at:  { type: string, format: date-time }
        last_used_at: { type: string, format: date-time, nullable: true }
...
security:
  - ApiKeyAuth: []

# ─────────────────────────────────────────────
# 路由定义
# ─────────────────────────────────────────────
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: 手机号注册（国内版）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, password, sms_code]
              properties:
                phone:    { type: string, example: "13800138000" }
                password: { type: string, minLength: 8 }
                sms_code: { type: string, example: "123456" }
      responses:
        "201":
          description: 注册成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: { type: string, format: uuid }
                  message: { type: string, example: "注册成功" }
        "400":
          description: 参数错误或验证码无效
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: 手机号已注册
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: 登录，获取 Bearer token（用于后续所有请求）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, password]
              properties:
                phone:    { type: string }
                password: { type: string }
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:  { type: string, description: "JWT，有效期 7 天" }
                  token_type:    { type: string, example: "Bearer" }
        "401":
          description: 手机号或密码错误
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/auth/sms:
    post:
      tags: [Auth]
      summary: 发送短信验证码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string }
      responses:
        "200":
          description: 验证码已发送
        "429":
          description: 发送过于频繁

  # ── API Key 管理 ──────────────────────────────
  /api/keys:
    get:
      tags: [API Keys]
      summary: 获取当前用户的所有 API Key
      responses:
        "200":
          description: Key 列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys: { type: array, items: { $ref: "#/components/schemas/ApiKey" } }

    post:
      tags: [API Keys]
      summary: 创建新的 API Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, example: "ComfyUI 主力机" }
      responses:
        "201":
          description: 创建成功，**仅此一次**返回完整 Key 明文
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:    { type: string, example: "sk-seed-ab12cdef..." }
                  detail: { $ref: "#/components/schemas/ApiKey" }
  /api/keys/{id}:
    delete:
      tags: [API Keys]
      summary: 撤销 API Key
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: 撤销成功 }
        "404": { description: Key 不存在 }

  # ── 余额 & 用量 ────────────────────────────────
...